A pull system means never assigning
This introduces a new communications section for describing how best to
communicate like a thoughtbotter.

Up first: tickets. We use Trello right now, but the principles of how we
use it are the same as when we used Pivotal Tracker, JIRA, Trajectory,
GitHub Issues, and everything else. Therefore I've left it as just
"tickets."

We follow a pull system, which breaks down like this:

* For client projects, the next designer or developer takes the ticket
  with top priority.
* For discussion topics, such as happens on the Research board, people
  subscribe to tickets they are interested in.

Assigning people to a ticket defeats this and causes confusion.
Therefore, we don't do that.