Simplify wording
We use a modified Kanban/CONWIP at thoughtbot. This commit message is
not the place to go into the full details of why we chose to do that,
but at its simplest the output is controlled by demand (e.g. "the
user wants this" or "the blog post author wants to write this") and
not by forecast ("this is due on Thursday", or by analogy "I want you to
write this"). This reflects how we believe software can be built and
how to run a sustainable, happy workplace.